Separate MetaMaskButton into a file - And rename MetaMaskButton to WalletSelectionButton
